# docs/swagger.yaml
definitions:
...
  dto.Category:
    properties:
      id:
        type: integer
      name:
        type: string
      store_id:
        type: integer
    type: object
  dto.CreateCategoryRequest:
    properties:
      name:
        type: string
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      total:
        type: integer
    type: object
...
  errorx.AppError:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        $ref: '#/definitions/errorx.ErrorCode'
      message:
        type: string
    type: object
  errorx.ErrorCode:
    enum:
    - UNAUTHORIZED
    - FORBIDDEN
    - VALIDATION_ERROR
    - NOT_FOUND
    - INTERNAL_ERROR
    - BAD_REQUEST
    - CONFLICT
    type: string
    x-enum-varnames:
    - ErrUnauthorized
    - ErrForbidden
    - ErrValidation
    - ErrNotFound
    - ErrInternal
    - ErrBadRequest
    - ErrConflict
  response.BaseResponse:
    properties:
      data: {}
      error: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /categories:
    get:
      description: Retrieve a paginated list of all global categories
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: List all categories
      tags:
      - Category
...
  /store/:store_id/categories:
    get:
      description: Retrieve all categories specific to a store
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Category'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: List store categories
      tags:
      - Category
  /store/:store_id/category:
    post:
      consumes:
      - application/json
      description: Creates a new custom category for a store
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      - description: Category Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Add store category
      tags:
      - Category
...
swagger: "2.0"